+++ src/components/Dashboard/DashboardBorrowSection.tsx
@@ -1,14 +1,24 @@
 import React, { useContext } from "react";
 import Card from "../UI/Card";
 import { DashboardContext } from "../../context/dashboardContext";
-import BorrowRepaySubsection from './BorrowRepaySubsection';
+import BorrowRepaySubsection from "./BorrowRepaySubsection";
 
 const DashboardBorrowSection = () => {
-  const { onPage } = useContext(DashboardContext);
+  const { onPage, repayProcessState } = useContext(DashboardContext);
   const subSection = onPage.split("-")[1];
+  let goBack: null | Function = null;
+
+  if (onPage === "Borrow-Repay") {
+    const { selectedLoan, setSelectedLoan } = repayProcessState;
+    if (selectedLoan) {
+      goBack = () => {
+        setSelectedLoan(null);
+      };
+    }
+  }
 
   return (
-    <Card className="main-card text-center" title={subSection}>
+    <Card className="main-card text-center" title={subSection} goBack={goBack}>
       <div className="my-5">
         {subSection == "Repay" && <BorrowRepaySubsection />}
         {subSection == "Deposit" && (
@@ -27,5 +37,3 @@ const DashboardBorrowSection = () => {
 };
 
 export default DashboardBorrowSection;
-
-
